pyproject.toml
[build-system]
requires = ["setuptools>=77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "fouroversix"
version = "0.1.0"
dependencies = [
    "torch~=2.9.1",
]

[project.optional-dependencies]
fast-build = [
    "ninja~=1.13.0",
    "psutil~=7.1.3",
]
test = [
    "lm_eval @ git+https://github.com/EleutherAI/lm-evaluation-harness.git",
    "modal~=1.2.4",
    "pytest~=9.0.1",
    "python-dateutil~=2.9.0.post0",
    "ruff~=0.14.7",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
indent-width = 4
exclude = ["build", "dist", "env", ".venv", "third_party"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
ignore = [
    # Various dumb rules related to docstrings
    "D100",
    "D103",  # TODO: Remove this one
    "D104",
    "D107",
    "D202",
    "D203",
    "D212",
    "FIX002",  # Allow TODO comments
    "PLC0415",  # Allow imports that aren't at the top of the file
    "PLR0913",  # Allow functions with more than 5 arguments
    "TD003",  # Don't require issue links on TODO comments
]
select = [
    "ALL",
]

[tool.ruff.lint.extend-per-file-ignores]
"src/fouroversix/ops.py" = [
    "I001",  # Allow unsorted imports, because torch needs to be imported before fouroversix._C
]
"src/fouroversix/quantize/triton_kernel.py" = [
    "ANN001",  # Allow missing annotations in Triton kernels
    "N803",  # Allow uppercase arguments to Triton kernels
    "N806",  # Allow uppercase variables in Triton kernels
]
"src/fouroversix/quantize/reference.py" = [
    "S101",  # Allow assert statements in reference implementation
]
"tests/**/test_*.py" = [
    "S101",  # Allow assert statements in tests
]
"scripts/**/*.py" = [
    "T201",  # Allow print statements in CLI tools
    "TID252",  # Allow relative imports from parent modules
]

[tool.ruff.lint.isort]
known-third-party = ["fouroversix"]